Prevent jumps to first line when scrolling popup
In older Vim versions pressing <c-j> or <c-k> will first jump to the
first line of the displayed buffer and start scrolling from there.
Running 'normal! zz' in the popup window seems to be a workaround in
some Vim versions to prevent this jump and start scrolling from the
current buffer line.

For more details see issue: https://github.com/vim/vim/issues/4876

Thanks to @nickspoons for the workaround.