Add new 'sign' option for highlighting error line
The sign entry can be set to highlight the error line in the displayed
buffer. Must be a dictionary specifying the sign attributes. See the
{dict} argument in sign_define() for more details.

When the 'number' column is enabled for the popup window and the 'text'
attribute has been set, the sign is always placed in the number column.