Use '\%v' for quickfix items with vcol = 1
When vcol is set to 1, we use '\%v' to match the display column.

Note 1: This doesn't work when the display column is past the last
character of the buffer line, `matchadd()` won't highlight anything.

TODO: We need to convert the display (or virtual/screen) columns to byte
columns, and then check whether they are within the buffer line, if not,
highlight the very last character of the buffer line.

Note 2: 'tabstop' needs to be set to 8, otherwise it doesn't highlight
the correct column (1 <tab> == 8).