Improve the popup column highlighting
In the previous commit `matchadd()` didn't highlight the quickfix
location in the popup window when the column position exceeded the
buffer line (sometimes happens with gcc, clang or clang-tidy output when
a brace is missing somewhere).

The following version first checks the length of the buffer line and, if
the column number exceeds the buffer line, highlights the last character
of that line.

When vcol = 1, we first calculate the byte index of the display column.