Don't call Get('sign') in popup callback
When the preview-popup is displayed and we hit <CR> in the quickfix
window, the cursor jumps to the quickfix locationa and _then_ the popup
is closed. This means, the popup callback is invoked after we left the
quickfix window. Get('sign') first checks b:qfpreview, and if it doesn't
exist, g:qfpreview. If we have set b:qfpreview in after/ftplugin/qf.vim,
instead of g:qfpreview, Get('sign') will always return false, and thus,
signs are not be removed. Therefore, remove the if-statement and just
call sign_unplace(). The function won't throw any errors if when no
signs have been placed.